+++ plugins/workload-management/src/main/java/org/opensearch/plugin/wlm/rule/action/CreateRuleRequest.java
@@ -10,13 +10,12 @@ package org.opensearch.plugin.wlm.rule.action;
 
 import org.opensearch.action.ActionRequest;
 import org.opensearch.action.ActionRequestValidationException;
-import org.opensearch.action.support.clustermanager.ClusterManagerNodeRequest;
+import org.opensearch.autotagging.Rule;
+import org.opensearch.autotagging.Rule.Builder;
 import org.opensearch.core.common.io.stream.StreamInput;
 import org.opensearch.core.common.io.stream.StreamOutput;
-import org.opensearch.core.common.io.stream.Writeable;
 import org.opensearch.core.xcontent.XContentParser;
-import org.opensearch.wlm.Rule;
-import org.opensearch.wlm.Rule.Builder;
+import org.opensearch.plugin.wlm.rule.QueryGroupFeatureType;
 import org.joda.time.Instant;
 
 import java.io.IOException;
@@ -26,13 +25,13 @@ import java.io.IOException;
  * @opensearch.experimental
  */
 public class CreateRuleRequest extends ActionRequest {
-    private final Rule rule;
+    private final Rule<QueryGroupFeatureType> rule;
 
     /**
      * Constructor for CreateRuleRequest
      * @param rule - A {@link Rule} object
      */
-    CreateRuleRequest(Rule rule) {
+    CreateRuleRequest(Rule<QueryGroupFeatureType> rule) {
         this.rule = rule;
     }
 
@@ -42,7 +41,7 @@ public class CreateRuleRequest extends ActionRequest {
      */
     CreateRuleRequest(StreamInput in) throws IOException {
         super(in);
-        rule = new Rule(in);
+        rule = new Rule<>(in);
     }
 
     /**
@@ -50,7 +49,7 @@ public class CreateRuleRequest extends ActionRequest {
      * @param parser - A {@link XContentParser} object
      */
     public static CreateRuleRequest fromXContent(XContentParser parser) throws IOException {
-        Builder builder = Builder.fromXContent(parser);
+        Builder<QueryGroupFeatureType> builder = Builder.fromXContent(parser, QueryGroupFeatureType.INSTANCE);
         return new CreateRuleRequest(builder.updatedAt(Instant.now().toString()).build());
     }
 
@@ -68,7 +67,7 @@ public class CreateRuleRequest extends ActionRequest {
     /**
      * Rule getter
      */
-    public Rule getRule() {
+    public Rule<QueryGroupFeatureType> getRule() {
         return rule;
     }
 }